TMUX use standard C-b prefix
The prefix C-a is making it hard to jump to the beginning of a line
(readline) or increment numerical values (Vim) -- actions which also use
this key combination. C-b is far less common and should be preferred.
Also, it's the standard prefix in tmux and will make it easier to use
tmux with the default config.

NOTE TO SELF: you should really NOT revert this commit. You are already
convinced C-b is the better choice, just stick with it! It will take
some getting used to, but you can make it!